requirements.txt
fastapi==0.115.13
pandas==2.2.3
beautifulsoup4==4.13.3
requests==2.32.3
flake8==7.3.0
uvicorn==0.34.3
streamlit==1.46.1
SQLAlchemy>=2.0,<3.0
alembic==1.16.4
pydantic[email]==2.2.0
PyJWT==2.10.1